Cleanup and renaming, as these things will be reused for more than openiddict in the future
